% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\docType{data}
\name{circa_table_genescarab}
\alias{circa_table_genescarab}
\title{plants}
\format{
Data frame with phase information in two different conditions for Marchantia polymorpha
}
\usage{
circa_table_genescarab
}
\description{
KEGG pathways associated to plants
}
\keyword{datasets}
