% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{circular_boxplot}
\alias{circular_boxplot}
\title{Circular boxplot}
\usage{
circular_boxplot(res.table, color.palette = "Tam", tr.height = 0.1)
}
\arguments{
\item{res.table}{A circular results table, as the output from
complete_circular_table or complete_circular_table_grouped}

\item{color.palette}{Color palette to use. Available palettes
are those from MetBrewer}

\item{tr.height}{Height of each track. For more than 8
tracks, reduce this parameter}
}
\value{
A PNG file containing the plot
}
\description{
Function to plot a circular boxplot showing the circular
distribution of different processes or gene sets
}
\examples{
org_Otaurireduced_eg_db <- load_example_annot()
data("circa_table_genescarab")
total_phases_table_sd <- data.frame(
    names = rownames(circa_table_genescarab),
    phase = as.numeric(
        circa_table_genescarab[["sd.peak.time.hours"]]
    )
)

functional_data <- AnnotationDbi::select(org_Otaurireduced_eg_db,
    keys = AnnotationDbi::keys(org_Otaurireduced_eg_db, keytype = "GID"),
    columns = c("GID", "GO")
)
complete_gos <- unique(functional_data$GO)
complete_gos <- complete_gos[!is.na(complete_gos)]

set.seed(2345)
subset_gos <- complete_gos[sample(
    seq_len(length(complete_gos)), 50,
    replace = FALSE
)]
go.list.test <- create_gene_list_go(
    go_vector = subset_gos,
    org.package = org_Otaurireduced_eg_db,
    go_column = "GO", id_column = "GID"
)
phases.list.sd <- gene_list_to_phases(
    go.list.test,
    total_phases_table_sd
)
phases.list.sd.clean <- phases.list.sd[which(
    sapply(phases.list.sd, nrow) != 0
)][seq_len(3)]
go.circa.table.sd <- complete_circular_table(
    phase.list = phases.list.sd.clean,
    hr.on.large.sets.th = 200,
    force.hr.th = 0.04,
    rao.on.large.sets.th = 200,
    force.rao.th = 0.04,
    iter.rao = 999, iter.hr = 999
)
circular_boxplot(go.circa.table.sd[seq_len(2), ],
    color.palette = "Tam"
)

}
