% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{complete_circular_table}
\alias{complete_circular_table}
\title{Creation of the circular table and deviation
from the uniform distribution for a phases list}
\usage{
complete_circular_table(
  phase.list,
  hr.on.large.sets.th = 400,
  iter.hr = 999,
  force.hr.th = 0.05,
  rao.on.large.sets.th = 400,
  iter.rao = 999,
  force.rao.th = 0.05
)
}
\arguments{
\item{phase.list}{Phases list in the same format as
the output of gene_list_to_phases}

\item{hr.on.large.sets.th}{Threshold indicating the
maximum size of a set to perform HR test on}

\item{iter.hr}{Number of iterations for HR test}

\item{force.hr.th}{Significance level to execute HR
test if Rayleigh and Kuiper p-values are higher}

\item{rao.on.large.sets.th}{Threshold indicating the
maximum size of a set to perform Rao test on}

\item{iter.rao}{Number of iterations for Rao test}

\item{force.rao.th}{Significance level to execute Rao
test if Rayleigh, Kuiper and HR p-values are higher}
}
\value{
A data.frame indicating, for each process or
gene set: number of phases used for computation (n); mean,
rho, variance, first, second and third quantiles of the
circular distribution in hours; p-values and adjusted
p-values for each of the four tests
}
\description{
Function that takes a list of continuous phases
associated with each process or gene set (output of the
gene_list_to_phases function) and calculates the
table of statistical measures and deviations from circular uniformity
for each one using the Rayleigh, Kuiper, Hermans-Rasson, and Rao tests.
}
\examples{
org_Otaurireduced_eg_db <- load_example_annot()
data("circa_table_genescarab")
total_phases_table_sd <- data.frame(
    names = rownames(circa_table_genescarab),
    phase = as.numeric(circa_table_genescarab[["sd.peak.time.hours"]])
)
functional_data <- AnnotationDbi::select(org_Otaurireduced_eg_db,
    keys = AnnotationDbi::keys(org_Otaurireduced_eg_db, keytype = "GID"),
    columns = c("GID", "GO")
)
complete_gos <- unique(functional_data$GO)
complete_gos <- complete_gos[!is.na(complete_gos)]

set.seed(2345)
subset_gos <- complete_gos[sample(
    seq_len(length(complete_gos)), 50,
    replace = FALSE
)]
go.list.test <- create_gene_list_go(
    go_vector = subset_gos,
    org.package = org_Otaurireduced_eg_db,
    go_column = "GO", id_column = "GID"
)
phases.list.sd <- gene_list_to_phases(
    go.list.test, total_phases_table_sd
)
phases.list.sd.clean <- phases.list.sd[which(
    sapply(phases.list.sd, nrow) != 0
)][seq_len(3)]
go.circa.table.sd <- complete_circular_table(
    phase.list = phases.list.sd.clean,
    hr.on.large.sets.th = 200,
    force.hr.th = 0.04,
    rao.on.large.sets.th = 200,
    force.rao.th = 0.04,
    iter.rao = 999, iter.hr = 999
)

}
