% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{create_gene_list_kegg}
\alias{create_gene_list_kegg}
\title{Creation of input lists for KEGG pathways from an annotation package}
\usage{
create_gene_list_kegg(
  ko_vector = c("all"),
  org.package,
  ko_column,
  id_column,
  ko_prefix = "map",
  species = "all"
)
}
\arguments{
\item{ko_vector}{KO vector to be included in the analysis,
along with their pathways ancestors. "all" takes the
complete set from the annotation package}

\item{org.package}{OrgDb object from species annotation package
(*.db)}

\item{ko_column}{KO column name in the annotation package}

\item{id_column}{Gene ID column name in the annotation package}

\item{ko_prefix}{Prefix used to denote KEGG pathways, either
"map" or "ko", depending on package version}

\item{species}{Species group to filter final pathways ("all"
for no filtering), either "bacteria", "archaea", "animals",
"plants", "fungi" or "protists"}
}
\value{
A list linking each KEGG pathway with the genes
that share that annotation
}
\description{
Function to create a list of genes associated with each KEGG
pathway based on a species annotation package with KO terms
derived from KEGGREST, KAAS or similar tools. It also allows
filtering by species group.
}
\examples{
org_Otaurireduced_eg_db <- load_example_annot()
create_gene_list_kegg(
    ko_vector = "K10666", org.package = org_Otaurireduced_eg_db,
    ko_column = "KO", id_column = "GID", ko_prefix = "map",
    species = "plants"
)

}
