% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.create_circular_table_grouped}
\alias{.create_circular_table_grouped}
\title{Creation of the circular table and deviation from the
uniform distribution for a single set of grouped data}
\usage{
.create_circular_table_grouped(
  circa_vector,
  n_bins,
  hr.on.large.sets.th = 400,
  iter.hr = 999,
  force.hr.th = 0.05,
  rao.on.large.sets.th = 400,
  iter.rao = 999,
  force.rao.th = 0.05,
  kuiper.on.large.sets.th = 400,
  iter.kuiper = 999,
  force.kuiper.th = 0.05
)
}
\arguments{
\item{circa_vector}{Discrete phases vector in hours}

\item{n_bins}{Number of bins}

\item{hr.on.large.sets.th}{Threshold indicating the maximum
size of a set to perform HR test on}

\item{iter.hr}{Number of iterations for HR test}

\item{force.hr.th}{Significance level to execute HR test if
Rayleigh and Kuiper p-values are higher}

\item{rao.on.large.sets.th}{Threshold indicating the maximum
size of a set to perform Rao test on}

\item{iter.rao}{Number of iterations for Rao test}

\item{force.rao.th}{Significance level to execute Rao test if
Rayleigh, Kuiper and HR p-values are higher}

\item{kuiper.on.large.sets.th}{Threshold indicating the maximum
size of a set to perform Kuiper test on}

\item{iter.kuiper}{Number of iterations for Kuiper test}

\item{force.kuiper.th}{Significance level to execute Kuiper
test if Rayleigh p-value is higher}
}
\value{
A data.frame indicating number of phases used for
computation (n); mean, rho, variance, first, second and third
quartiles of the circular distribution in hours; and
p-values for each of the four tests
}
\description{
Function that takes a vector of gene phases (discrete
measurements or non-parametric estimates)
and returns statistical measures of their circular
distribution and their deviations from a circular uniform
distribution based on Rayleigh and adapted versions of
Kuiper, Hermans-Rasson and Rao tests.
}
