% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.get_ancestor_for_gene_lists}
\alias{.get_ancestor_for_gene_lists}
\title{Extraction of all ancestors of a given GO terms vector}
\usage{
.get_ancestor_for_gene_lists(go_vector)
}
\arguments{
\item{go_vector}{GO terms vector to be included in the
analysis, along with their ancestors. "all" takes the
complete set from the annotation package}
}
\value{
A vector containing ancestors of GO terms
}
\description{
Function to extract all ancestors of given GO terms
}
