% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.HermansRasson2PGroupedRad}
\alias{.HermansRasson2PGroupedRad}
\title{HR test's p-value calculation for grouped data}
\usage{
.HermansRasson2PGroupedRad(sample, m, iter = 9999)
}
\arguments{
\item{sample}{Data points in radians}

\item{m}{Number of bins}

\item{iter}{Number of iterations}
}
\value{
The p-value for the test
}
\description{
Helper function that calculates the p-value of an adaptation of the
Hermans-Rasson test for grouped data (discrete phases) as
implemented in the article "Grouped circular data in biology:
advice for effectively implementing statistical procedures"
https://doi.org/10.1007/s00265-020-02881-6.
}
