% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.HermansRasson2T}
\alias{.HermansRasson2T}
\title{Critical value calculation for HR test}
\usage{
.HermansRasson2T(sample)
}
\arguments{
\item{sample}{Data points in radians}
}
\value{
The critical value for the test
}
\description{
A helper function that calculates the critical value for the HR test,
as implemented in the article "Grouped circular data
in biology: advice for effectively implementing statistical procedures"
https://doi.org/10.1007/s00265-020-02881-6.
}
