% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.input_create_circular_table}
\alias{.input_create_circular_table}
\title{Input checking for .create_circular_table and
.create_circular_table_grouped}
\usage{
.input_create_circular_table(circa_vector)
}
\arguments{
\item{circa_vector}{Phases vector in hours}
}
\value{
An error if input is not adequate
}
\description{
Helper function checking correctness of input for
.create_circular_table and .create_circular_table_grouped
}
