% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.input_create_gene_list_go}
\alias{.input_create_gene_list_go}
\title{Input checking for create_gene_list_go}
\usage{
.input_create_gene_list_go(go_vector, org.package, go_column, id_column)
}
\arguments{
\item{go_vector}{GO terms vector to be included in the
analysis, along with their ancestors. "all" takes the
complete set from the annotation package}

\item{org.package}{OrgDb object from species annotation package
(*.db)}

\item{go_column}{GO term column name in the annotation package}

\item{id_column}{Gene ID column name in the annotation package}
}
\value{
A data.frame showing GO terms per gene as appear
in the annotation package
}
\description{
Helper function checking correctness of input for
create_gene_list_go and generating a data.frame with
all the GO terms per gene
}
