% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.input_create_gene_list_kegg}
\alias{.input_create_gene_list_kegg}
\title{Input checking for create_gene_list_kegg}
\usage{
.input_create_gene_list_kegg(
  ko_vector = c("all"),
  org.package,
  ko_column,
  id_column,
  ko_prefix = "map",
  species = "all"
)
}
\arguments{
\item{ko_vector}{KO vector to be included in the analysis,
along with their pathways ancestors. "all" takes the
complete set from the annotation package}

\item{org.package}{OrgDb object from species annotation package
(*.db)}

\item{ko_column}{KO column name in the annotation package}

\item{id_column}{Gene ID column name in the annotation package}

\item{ko_prefix}{Prefix used to denote KEGG pathways, either
"map" or "ko", depending on package version}

\item{species}{Species group to filter final pathways ("all"
for no filtering), either "bacteria", "archaea", "animals",
"plants", "fungi" or "protists"}
}
\value{
A data.frame showing KO terms per gene as appear
in the annotation package
}
\description{
Helper function checking correctness of input for
create_gene_list_kegg and generating a data.frame with
all the KOs per gene
}
