% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.input_gene_contribution}
\alias{.input_gene_contribution}
\title{Input checking for gene_contribution_to_set}
\usage{
.input_gene_contribution(circa_result, phase_list)
}
\arguments{
\item{circa_result}{Table with circular distribution
results, as the output from
complete_circular_table or complete_circular_table_grouped}

\item{phase_list}{Phases list in the same format as
the output of gene_list_to_phases, containing the row
names of circa_result as names of each element}
}
\value{
A phase list containing only elements that appear in
circa_result rows
}
\description{
Helper function checking correctness of input for
gene_contribution_to_set and returning a clean phase list
}
