% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.legend_builder}
\alias{.legend_builder}
\title{Helper legend builder for dotplot and histogram}
\usage{
.legend_builder(phases.list, my_color)
}
\arguments{
\item{phases.list}{A named phase tables list in the same
format as the output of gene_list_to_phases}

\item{my_color}{Color vector to use}
}
\value{
Legend of the plots
}
\description{
Function to create tracks for dotplot
}
