% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.legend_builder_boxplot}
\alias{.legend_builder_boxplot}
\title{Helper legend builder for boxplot}
\usage{
.legend_builder_boxplot(quan.table, my_color)
}
\arguments{
\item{quan.table}{A circular results table with sem info}

\item{my_color}{Color vector to use}
}
\value{
Legend of the plots
}
\description{
Function to create tracks for boxplot
}
