% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.pathways_from_kos}
\alias{.pathways_from_kos}
\title{Extraction of KEGG pathways from KO terms}
\usage{
.pathways_from_kos(ko_vector, ko_prefix)
}
\arguments{
\item{ko_vector}{KO vector to extract pathways}

\item{ko_prefix}{Prefix used to denote KEGG pathways, either
"map" or "ko", depending on package version}
}
\value{
A list of pathways associated to the KO
terms input
}
\description{
Function to extract KEGG pathways associated to a
vector of KO terms and create a list containing this
information
}
