% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.RaoTestUngroupedRad}
\alias{.RaoTestUngroupedRad}
\title{Rao's spacing test's p-value calculation for ungrouped data}
\usage{
.RaoTestUngroupedRad(sample, iter = 9999)
}
\arguments{
\item{sample}{Data points in radians}

\item{iter}{Number of iterations}
}
\value{
The p-value for the test
}
\description{
Helper function that calculates the p-value of Rao's spacing test
for ungrouped data (continuous phases) as implemented
in the article "Circular statistics meets practical
limitations: a simulation-based Rao’s spacing
test for non-continuous data"
https://doi.org/10.1186/s40462-019-0160-x.
}
