% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{.table_builder}
\alias{.table_builder}
\title{Helper function to construct table from circular data}
\usage{
.table_builder(
  circa_summary,
  circa_var,
  circa_kuiper,
  circa_ray,
  circa_hr,
  circa_rao
)
}
\arguments{
\item{circa_summary}{Process or gene sets list (as the output
of create_gene_list_go and create_gene_list_kegg)}

\item{circa_var}{Table showing acrophases for the
complete gene set under study, stated in hours}

\item{circa_kuiper}{P-value estimated by Kuiper test}

\item{circa_ray}{P-value estimated by Rayleigh test}

\item{circa_hr}{P-value estimated by HR test}

\item{circa_rao}{P-value estimated by Rao test}
}
\value{
A table indicating number of phases used for
computation (n); mean, rho, variance, first,
second and third quantiles of the circular distribution
in hours; and p-values for each of the four tests
}
\description{
Function that takes the calculated circular summary,
variance and p-values of Rayleigh, Kuiper, HR and Rao's tests and
construct the result table.
}
