% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{input_complete_circular_table}
\alias{input_complete_circular_table}
\title{Input checking for complete_circular_table and
complete_circular_table_grouped}
\usage{
input_complete_circular_table(phase.list)
}
\arguments{
\item{phase.list}{Phases list in the same format as
the output of gene_list_to_phases}
}
\value{
An error if input is not adequate
}
\description{
Helper function checking correctness of input for
complete_circular_table and complete_circular_table_grouped
}
