% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{load_example_annot}
\alias{load_example_annot}
\title{Loading annotation for examples and vignettes}
\usage{
load_example_annot()
}
\value{
OrgDb object
}
\description{
Loading SQlite containing GO and KO annnotation info for Ostreococcus
tauri in AnnotationForge's output format. Original genome and
annotation package were download from
https://github.com/fran-romero-campero/AlgaeFUN and KO information was
generated via KAAS (https://www.genome.jp/kaas-bin/kaas_main).
}
\examples{
org_Otaurireduced_eg_db <- load_example_annot()

}
