% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{multimodal_modes_test}
\alias{multimodal_modes_test}
\title{Identification of different modes}
\usage{
multimodal_modes_test(phase_table)
}
\arguments{
\item{phase_table}{A phase table, as the individual
elements of the list derived from gene_list_to_phases}
}
\value{
A vector containing the p-values for 1,2 and 3
f-fold symmetry
}
\description{
Function to determine the extract the Rayleigh p-value
in the presence of f-fold symmetry (1, 2 or 3)
}
