% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{RaoPGroupedRad}
\alias{RaoPGroupedRad}
\title{Rao's spacing test's p-value calculation for grouped data}
\usage{
RaoPGroupedRad(sample, m, k = 1000, iter = 9999)
}
\arguments{
\item{sample}{Data points in radians}

\item{m}{Number of bins}

\item{k}{Kappa for the error distribution}

\item{iter}{Number of iterations}
}
\value{
The p-value for the test
}
\description{
Helper function that calculates the p-value of an adaptation of the Rao's spacing test
for grouped data (discrete phases) as implemented in the article "Grouped circular data in biology: advice for effectively
implementing statistical procedures" https://doi.org/10.1007/s00265-020-02881-6.
}
