% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{track_box_box_cut}
\alias{track_box_box_cut}
\title{Helper track creation boxplot for box cutting the origin}
\usage{
track_box_box_cut(tr.height, quan.table, i, my_color)
}
\arguments{
\item{tr.height}{Height of each track. For more than 8
tracks, reduce this parameter}

\item{quan.table}{A circular results table with sem info}

\item{i}{Track counter}

\item{my_color}{Color vector to use}
}
\value{
Plotted track
}
\description{
Function to create tracks for boxplots whose boxes
intersect the origin. In this case, the segment must
be cut from min to 24 and from 0 to max. Also, the
box must be unfolded from first to 24 and from 0 to third
quartiles.
}
