% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{track_box_no_cut}
\alias{track_box_no_cut}
\title{Helper track creation boxplot without cut}
\usage{
track_box_no_cut(tr.height, quan.table, i, my_color)
}
\arguments{
\item{tr.height}{Height of each track. For more than 8
tracks, reduce this parameter}

\item{quan.table}{A circular results table with sem info}

\item{i}{Track counter}

\item{my_color}{Color vector to use}
}
\value{
Plotted track
}
\description{
Function to create tracks for boxplots that do not intersect
the origin. If neither of the segments cut the sunrise,
}
