% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{track_dot}
\alias{track_dot}
\title{Helper track creation dotplot}
\usage{
track_dot(tr.height, phases.list, i, my_color, my_transparent)
}
\arguments{
\item{tr.height}{Height of each track. For more than 8
tracks, reduce this parameter}

\item{phases.list}{A named phase tables list in the same
format as the output of gene_list_to_phases}

\item{i}{Track counter}

\item{my_color}{Color vector to use}

\item{my_transparent}{Transparent color vector to use}
}
\value{
Axis
}
\description{
Function to create tracks for dotplot
}
