% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\docType{data}
\name{circa_table_genescarab}
\alias{circa_table_genescarab}
\title{circa_table_genescarab}
\format{
Data frame
}
\usage{
circa_table_genescarab
}
\description{
Dataset example with phase information in two different conditions for Ostreococcus tauri
from "Multiomics integration unveils photoperiodic plasticity in the molecular rhythms of
marine phytoplankton" (https://doi.org/10.1093/plcell/koaf033)
}
\keyword{datasets}
