% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{circular_boxplot}
\alias{circular_boxplot}
\title{Circular boxplot}
\usage{
circular_boxplot(
  res.table,
  filename,
  color.palette = "Tam",
  tr.height = 0.1,
  width = 1600,
  height = 1600,
  res = 180
)
}
\arguments{
\item{res.table}{A circular results table, as the output from complete_circular_table or complete_circular_table_grouped}

\item{filename}{PNG file name to export the plot}

\item{color.palette}{Color palette to use. Available palettes are those from MetBrewer}

\item{tr.height}{Height of each track. For more than 8 tracks, reduce this parameter}

\item{width}{PNG file width in px}

\item{height}{PNG file height in px}

\item{res}{PNG file resolution}
}
\value{
A PNG file containing the plot
}
\description{
Function to plot a circular boxplot showing the circular distribution of
different processes or gene sets
}
