% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{circular_dotplot}
\alias{circular_dotplot}
\title{Circular dotplot}
\usage{
circular_dotplot(phases.list, color.palette = "Tam", tr.height = 0.1)
}
\arguments{
\item{phases.list}{A named phase tables list in the same
format as the output of gene_list_to_phases}

\item{color.palette}{Color palette to use. Available palettes
are those from MetBrewer}

\item{tr.height}{Height of each track. For more than 8 tracks,
reduce this parameter}
}
\value{
A PNG file containing the plot
}
\description{
Function to plot a circular dotplot showing the circular distribution of
different processes or gene sets
}
\examples{
library(org.Otauri.eg.db)
data("circa_table_genescarab")
total_phases_table_sd <- data.frame(
    names = rownames(circa_table_genescarab),
    phase = as.numeric(
        circa_table_genescarab[["sd.peak.time.hours"]]
    )
)

functional_data <- select(org.Otauri.eg.db,
    keys = keys(org.Otauri.eg.db, keytype = "GID"),
    columns = c("GID", "GO")
)
complete_gos <- unique(functional_data$GO)
complete_gos <- complete_gos[!is.na(complete_gos)]

set.seed(2345)
subset_gos <- complete_gos[sample(
    seq_len(length(complete_gos)), 50,
    replace = FALSE
)]
go.list.test <- create_gene_list_go(
    go_vector = subset_gos,
    org.package = "org.Otauri.eg.db",
    go_column = "GO", id_column = "GID"
)
phases.list.sd <- gene_list_to_phases(
    go.list.test, total_phases_table_sd
)
phases.list.sd.clean <- phases.list.sd[which(
    sapply(phases.list.sd, nrow) != 0
)][seq_len(3)]
circular_dotplot(phases.list.sd.clean, color.palette = "Tam")

}
