% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{circular_histogram}
\alias{circular_histogram}
\title{Circular histogram}
\usage{
circular_histogram(
  phases.list,
  filename,
  color.palette = "Tam",
  tr.height = 0.08,
  nbins = 24,
  width = 1600,
  height = 1600,
  res = 180
)
}
\arguments{
\item{phases.list}{A phase tables list in the same format as the output of gene.list.to.phases}

\item{filename}{PNG file name to export the plot}

\item{color.palette}{Color palette to use. Available palettes are those from MetBrewer}

\item{tr.height}{Height of each track. For more than 8 tracks, reduce this parameter}

\item{nbins}{Number of bins (breaks) to split the histogram}

\item{width}{PNG file width in px}

\item{height}{PNG file height in px}

\item{res}{PNG file resolution}
}
\value{
A PNG file containing the plot
}
\description{
Function to plot a circular histogram showing the circular distribution of
different processes or gene sets
}
