% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{complete_circular_table_grouped}
\alias{complete_circular_table_grouped}
\title{Creation of the circular table and deviation from the uniform distribution for a discrete (grouped) phases list}
\usage{
complete_circular_table_grouped(
  phase.list,
  n_bins,
  hr.on.large.sets.th = 400,
  iter.hr = 999,
  force.hr.th = 0.05,
  rao.on.large.sets.th = 400,
  iter.rao = 999,
  force.rao.th = 0.05,
  kuiper.on.large.sets.th = 400,
  iter.kuiper = 999,
  force.kuiper.th = 0.05
)
}
\arguments{
\item{phase.list}{Phases list in the same format as the output of gene.list.to.phases}

\item{n_bins}{Number of bins}

\item{hr.on.large.sets.th}{Threshold indicating the maximum size of a set to perform HR test on}

\item{iter.hr}{Number of iterations for HR test}

\item{force.hr.th}{Significance level to execute HR test if Rayleigh and Kuiper p-values are higher}

\item{rao.on.large.sets.th}{Threshold indicating the maximum size of a set to perform Rao test on}

\item{iter.rao}{Number of iterations for Rao test}

\item{force.rao.th}{Significance level to execute Rao test if Rayleigh, Kuiper and HR p-values are higher}

\item{kuiper.on.large.sets.th}{Threshold indicating the maximum size of a set to perform kuiper test on}

\item{iter.kuiper}{Number of iterations for kuiper test}

\item{force.kuiper.th}{Significance level to execute Kuiper test if Rayleigh p-value is higher}
}
\value{
A data.frame indicating, for each process or gene set: number of phases used for computation (n); mean, rho, variance, first, second and third quantiles of the circular distribution in hours; p-values and adjusted p-values for each of the four tests
}
\description{
Function that takes a list of discrete phases associated with each process or gene set (output of the
gene.list.to.phases function) and calculates the table of statistical measures and deviations from circular uniformity
for each one using the adapted Rayleigh, Kuiper, Hermans-Rasson, and Rao tests.
}
