% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{gene_contribution_to_set}
\alias{gene_contribution_to_set}
\title{Contribution of each gene in a set to the temporal cohesion of the whole set}
\usage{
gene_contribution_to_set(circa_result, phase_list)
}
\arguments{
\item{circa_result}{Table with circular distribution results, as the output from complete_circular_table or complete_circular_table_grouped}

\item{phase_list}{Phases list in the same format as the output of gene.list.to.phases, containing the row names of circa_result as names of each element}
}
\value{
A list whose elements are tables indicating the difference in circular mean and rho due to removing each gene for each process or gene set, and the rank of each difference
}
\description{
Function that measures the contribution to the mean and rho of the circular distribution of a
process or gene set for each of its constituent genes by removing them one by one. For sets
consisting on more than 100 genes, it takes a random sample of 100 gene phases to normalize
and ensure noticiable changes.
}
