% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\docType{PACKAGE}
\name{GeneSCARAB}
\alias{GeneSCARAB}
\title{GeneSCARAB: Gene Set Circular Analysis and Rhythms Ascertainment in
Bioprocesses}
\description{
GeneSCARAB provides methods for analyzing the cohesion and clustering toward
specific points in the cycle of biological sets (genes, metabolites, ...)
for chronobiology studies following the proposed best practices in circular
statistics.
}
\details{
\subsection{Main workflow}{

The typical workflow is:
\enumerate{
\item Load element phases as a table and species annotation package.
\item Construct the phases list for the desired sets or processes.
\item Compute circular statistics and departure from uniformity for
each set.
\item For non-uniform sets, compare against global phase distribution.
\item In the case of multiple conditions, determine differentially timed sets
between conditions.
\item For multimodal sets, analyze the number of modes and cluster its elements.
\item Calculate contribution of each element to its set's statistics.
\item Plot results as circular boxplot, dotplot or histogram.
}
}

\subsection{Main functions}{
\itemize{
\item \code{\link[=create_gene_list_go]{create_gene_list_go()}} creates a list of elements associated with each GO
term based on a species annotation package.
\item \code{\link[=create_gene_list_kegg]{create_gene_list_kegg()}} create a list of elements associated with each
KEGG pathway based on a species annotation package.
\item \code{\link[=gene_list_to_phases]{gene_list_to_phases()}} takes a list of elements associated with
specific sets and a complete phase table, and returns a list
consisting of the individual phase tables of each set.
\item \code{\link[=complete_circular_table]{complete_circular_table()}} calculates the table of statistical measures
and deviations from circular uniformity of each set.
\item \code{\link[=complete_circular_table_grouped]{complete_circular_table_grouped()}} calculates the table of statistical
measures and deviations from circular uniformity of each set (for
discrete phases).
\item \code{\link[=test_against_gen_dist]{test_against_gen_dist()}} compares whether the distributions
of specific sets significantly differ from the overall phase distribution
of the entire dataset
\item \code{\link[=test_two_dist]{test_two_dist()}} compares whether the distributions of specific
sets differ significantly between two different conditions or genotypes.
\item \code{\link[=gene_contribution_to_set]{gene_contribution_to_set()}} measures the contribution of each element
of the set to the mean and rho of the entire set.
\item \code{\link[=multimodal_analysis]{multimodal_analysis()}} determines the most likely number of modes
in the presence of f-fold symmetry (1, 2 or 3),
cluster elements in f clusters and determine the circular statistic
summary for each cluster
\item \code{\link[=circular_boxplot]{circular_boxplot()}} creates circular boxplot.
\item \code{\link[=circular_dotplot]{circular_dotplot()}} creates circular dotplot.
\item \code{\link[=circular_histogram]{circular_histogram()}} creates circular histogram.
}

See the package vignette for a complete tutorial.
}
}
\keyword{circadian,}
\keyword{enrichment,}
\keyword{omics,}
\keyword{phases}
\keyword{rhythms,}
