% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{multimodal_analysis}
\alias{multimodal_analysis}
\title{Multimodality analysis}
\usage{
multimodal_analysis(circa_table)
}
\arguments{
\item{circa_table}{A phase table, as the individual elements of the list derived from gene.list.to.phases}
}
\value{
A list containing the most likely mode number, the p-value for that number, the p-values corresponding to each mode number, the circular summary table for each cluster and the genes contained in each cluster
}
\description{
Function to determine the most likely number of modes in the presence of f-fold symmetry (1, 2 or 3),
cluster genes in f clusters and determine the statistic summary for each cluster
}
