% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{test_against_gen_dist}
\alias{test_against_gen_dist}
\title{Deviation from the total experimental distribution for a phases list}
\usage{
test_against_gen_dist(phase.list, total.phase.table)
}
\arguments{
\item{phase.list}{Phases list in the same format as the output of gene.list.to.phases}

\item{total.phase.table}{Table showing acrophases for the complete gene set under study, stated in hours}
}
\value{
A data.frame with p-value and adjusted p-value (null hyphotesis: same distribution as the general one) for each process or gene set
}
\description{
Function that compares whether the distributions of specific processes or gene sets differ
significantly from the overall phase distribution of the entire study. Significance is calculated
based on a MANOVA analysis as described in https://doi.org/10.1186/s40462-022-00323-8.
}
