% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{test_against_gen_dist}
\alias{test_against_gen_dist}
\title{Deviation from the total experimental distribution for a phases list}
\usage{
test_against_gen_dist(phase.list, total.phase.table)
}
\arguments{
\item{phase.list}{Phases list in the same format as the output of gene_list_to_phases}

\item{total.phase.table}{Table showing acrophases for the complete gene set under study, stated in hours}
}
\value{
A data.frame with p-value and adjusted p-value (null hyphotesis: same distribution as the general one) for each process or gene set
}
\description{
Function that compares whether the distributions of specific processes or gene sets differ
significantly from the overall phase distribution of the entire study. Significance is calculated
based on a MANOVA analysis as described in https://doi.org/10.1186/s40462-022-00323-8.
}
\examples{
library(org.Otauri.eg.db)
data("circa_table_genescarab")
total_phases_table_sd <- data.frame(names=rownames(circa_table_genescarab),
phase=as.numeric(circa_table_genescarab[["sd.peak.time.hours"]]))

functional_data <- select(org.Otauri.eg.db,
                          keys = keys(org.Otauri.eg.db,keytype="GID"),
                          columns =  c("GID", "GO"))
complete_gos <- unique(functional_data$GO)
complete_gos <- complete_gos[!is.na(complete_gos)]

set.seed(2345)
subset_gos <- complete_gos[sample(1:length(complete_gos), 50, replace = FALSE)]
go.list.test <- create_gene_list_go(go_vector = subset_gos,
org.package = "org.Otauri.eg.db",
go_column = "GO", id_column = "GID")
phases.list.sd <- gene_list_to_phases(go.list.test, total_phases_table_sd)
phases.list.sd.clean <- phases.list.sd[which(sapply(phases.list.sd, nrow) != 0)][1:3]
not_experimentally_distributed_gos <- test_against_gen_dist(phases.list.sd.clean,
total_phases_table_sd)

}
