% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{test_two_dist}
\alias{test_two_dist}
\title{Comparison of circular distributions between two conditions}
\usage{
test_two_dist(phase.list.1, phase.list.2)
}
\arguments{
\item{phase.list.1}{Phases list for condition 1, in the same format as the output of gene.list.to.phases}

\item{phase.list.2}{Phases list for condition 2, in the same format as the output of gene.list.to.phases}
}
\value{
A data.frame with p-value and adjusted p-value (null hyphotesis: same distribution in both conditions) for each process or gene set
}
\description{
Function that compares whether the distributions of specific processes or gene sets differ
significantly between two different conditions or genotypes. The two lists must have the same
number of elements, although the number of genes associated with each process or gene set may differ.
Significance is calculated based on a MANOVA analysis as described in https://doi.org/10.1186/s40462-022-00323-8.
}
