% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genescarab.R
\name{test_two_dist}
\alias{test_two_dist}
\title{Comparison of circular distributions between two conditions}
\usage{
test_two_dist(phase.list.1, phase.list.2)
}
\arguments{
\item{phase.list.1}{Phases list for condition 1, in the same
format as the output of gene_list_to_phases}

\item{phase.list.2}{Phases list for condition 2, in the same
format as the output of gene_list_to_phases}
}
\value{
A data.frame with p-value and adjusted p-value (null
hyphotesis: same distribution in both conditions) for each
process or gene set
}
\description{
Function that compares whether the distributions of specific
processes or gene sets differ
significantly between two different conditions or genotypes.
The two lists must have the same
number of elements, although the number of genes associated
with each process or gene set may differ.
Significance is calculated based on a MANOVA analysis as
described in https://doi.org/10.1186/s40462-022-00323-8.
}
\examples{
library(org.Otauri.eg.db)
data("circa_table_genescarab")
total_phases_table_sd <- data.frame(
    names = rownames(circa_table_genescarab),
    phase = as.numeric(
        circa_table_genescarab[["sd.peak.time.hours"]]
    )
)
total_phases_table_ld <- data.frame(
    names = rownames(circa_table_genescarab),
    phase = as.numeric(
        circa_table_genescarab[["ld.peak.time.hours"]]
    )
)

functional_data <- select(org.Otauri.eg.db,
    keys = keys(org.Otauri.eg.db, keytype = "GID"),
    columns = c("GID", "GO")
)
complete_gos <- unique(functional_data$GO)
complete_gos <- complete_gos[!is.na(complete_gos)]

set.seed(2345)
subset_gos <- complete_gos[sample(
    seq_len(length(complete_gos)), 50,
    replace = FALSE
)]
go.list.test <- create_gene_list_go(
    go_vector = subset_gos,
    org.package = "org.Otauri.eg.db",
    go_column = "GO", id_column = "GID"
)
phases.list.sd <- gene_list_to_phases(
    go.list.test, total_phases_table_sd
)
phases.list.sd.clean <- phases.list.sd[which(
    sapply(phases.list.sd, nrow) != 0
)][seq_len(3)]

phases.list.ld <- gene_list_to_phases(
    go.list.test, total_phases_table_sd
)
phases.list.ld.clean <- phases.list.sd[which(
    sapply(phases.list.sd, nrow) != 0
)][names(phases.list.sd.clean)]


diff_distributed_gos <- test_two_dist(
    phases.list.ld.clean, phases.list.sd.clean
)

}
